# IDGS901_API_Balones/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BalonEvo",
    "version": "v1"
  },
  "paths": {
...
    "/api/Pedidos": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedidos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedidos/insertar_detalle": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DetallePedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedidos/{usuario}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedidos/verDetallePedido": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Pedidos/actualizarEnProcesoAEnCamino/{idPedido}": {
      "put": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idPedido",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedidos/actualizarEnCaminoAEntregado/{idPedido}": {
      "put": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idPedido",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "operationId": "Usuarios",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Clientes": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "rol": {
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
    "DetallePedido": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "pedidos": {
          "$ref": "#/definitions/Pedidos"
        },
        "productos": {
          "$ref": "#/definitions/Productos"
        }
      }
    },
    "Direccion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCompleto": {
          "type": "string"
        },
        "calleNumero": {
          "type": "string"
        },
        "codigoPostal": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "usuario": {
          "$ref": "#/definitions/Clientes"
        }
      }
    },
...
    "Pedidos": {
      "type": "object",
      "properties": {
        "idPedido": {
          "format": "int32",
          "type": "integer"
        },
        "cliente": {
          "$ref": "#/definitions/Clientes"
        },
        "direccion": {
          "$ref": "#/definitions/Direccion"
        },
        "tarjeta": {
          "$ref": "#/definitions/Tarjetas"
        },
        "folio": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
    "Productos": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
    "Tarjetas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTarjeta": {
          "type": "string"
        },
        "numTarjeta": {
          "type": "string"
        },
        "fechaVencimiento": {
          "type": "string"
        },
        "ccv": {
          "type": "string"
        },
        "usuario": {
          "$ref": "#/definitions/Clientes"
        }
      }
    },
    "Usuarios": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellidoPaterno": {
          "type": "string"
        },
        "apellidoMaterno": {
          "type": "string"
        },
        "edad": {
          "format": "int32",
          "type": "integer"
        },
        "sexo": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "rol": {
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
  }
}